+++ frontend/src/components/screens/dashboard/hhp/technicians/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const HHPTechniciansAnalyticsScreen = () => {
-    return (
-        <div>HHPTechniciansAnalytics</div>
-    )
-}
-
-export default HHPTechniciansAnalyticsScreen