+++ frontend/src/components/screens/dashboard/hhp/technicians/table.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const HHPDashboardTable = () => {
+    return (
+        <div>HHPDashboardTable</div>
+    )
+}
+
+export default HHPDashboardTable