+++ server/controllers/department/hhp/bin_dashboard/get_stats.js
@@ -2,19 +2,7 @@ import { pool } from "../../../../db.js";
 const GetBinStats = async (req, res) => {
     try {
         const { rows } = await pool.query(
-            //             `SELECT engineer, date_booked, unit_status, COUNT(*)::INTEGER AS units_count,
-            //        ARRAY_AGG(
-            //            jsonb_build_object('ticket_number', ticket_number, 'ticket_id', repairshopr_job_id)
-            //            ORDER BY date_booked DESC
-            //        ) AS tickets
-            // FROM technician_tasks
-            // WHERE unit_status IN ('New', 'In Progress', 'Customer Reply', 'Assigned to Tech',
-            //                       'Parts Request 1st Approval', 'Parts to be ordered', 'Waiting for Parts')
-            // GROUP BY engineer, unit_status
-            // ORDER BY engineer, unit_status;
-            // `
-            `
-SELECT 
+            `SELECT 
     engineer, 
     unit_status,
     COUNT(*)::INTEGER AS units_count, 
@@ -40,9 +28,6 @@ WHERE unit_status IN (
 ) 
 GROUP BY engineer, unit_status
 ORDER BY engineer, unit_status;
-
-
-
 `
         );