+++ server/routes/comments/index.js
@@ -6,8 +6,6 @@ import { authenticateToken } from "../../middleware/verify.js";
 
 const router = express.Router();
 
-// todo: remove and uncomment correct one
-router.post("/", addComment);
-// router.post("/", limiter, authenticateToken, addComment);
+router.post("/", limiter, authenticateToken, addComment);
 router.get("/", authenticateToken, getComments);
 export { router };