+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.31] - 2025-03-27
+
+### Added
+
+-   allow more faults in create ticket screen
+
+### Fixed
+
+-   bin stats calculations are glitching, calculate on the client
+-   group select by technicians in bin stats
+
 ## [0.1.30] - 2025-03-25
 
 ### Added